Fixed compiler warnings

+++ src/SpdReaderWriterDll/Driver/WinRing0.cs
@@ -487,7 +487,7 @@ namespace SpdReaderWriterDll.Driver {
             for (ushort bus = 0; bus <= gPciNumberOfBus; bus++) {
                 for (byte dev = 0; dev < gPciNumberOfDevice; dev++) {
 
-                    if (ReadPciConfigWord(PciBusDevFunc(bus, dev, 0), 0x00) == uint.MaxValue) {
+                    if (ReadPciConfigWord(PciBusDevFunc(bus, dev, 0), 0x00) == ushort.MaxValue) {
                         continue;
                     }
